Updated html to fit kintone
The HTML has been updated with new content, and has been modified to fit the customized view of kintone. It's not technically an html file, so it is not saved as html. This code needs to be copy and pasted in your Apps settings. Settings -> Create View -> Custom View -> HTML